backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
qdrant-client==1.7.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
pyyaml==6.0.1
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0
numpy>=1.24.0
sentence-transformers>=2.2.2
psutil>=5.9.0
